requirements.txt
replicate
python-dotenv
requests